First step in refactoring Java side
