+++ org.mwc.cmap.combined.feature/root_installs/sample_data/scripts/format_track.js
@@ -4,7 +4,8 @@
  */
 
 loadModule("/System/UI", false);
-loadModule("/Debrief", false);
+loadModule("/Debrief/Core", false);
+loadModule("/Debrief/Spatial", false);
 
 var selection = getSelection();
