+++ org.mwc.cmap.combined.feature/root_installs/sample_data/scripts/on_time_demo.js
@@ -4,3 +4,4 @@
  * onEventBus : info/debrief/newTime
  */
 
+print("new time received in JS");
