+++ org.mwc.debrief.scripting/plugin.xml
@@ -5,12 +5,33 @@
    <extension
          point="org.eclipse.ease.modules">
       <module
+            category="org.mwc.debrief.scripting.Debrief_Category"
             class="org.mwc.debrief.scripting.wrappers.DebriefObjects"
             icon="icons/debrief_icon.png"
             id="org.mwc.debrief.scripting.module"
-            name="Debrief"
+            name="Core"
             visible="true">
       </module>
+      <module
+            category="org.mwc.debrief.scripting.Debrief_Category"
+            class="org.mwc.debrief.scripting.wrappers.Shapes"
+            icon="icons/debrief_icon.png"
+            id="org.mwc.debrief.scripting.module"
+            name="Shapes"
+            visible="true">
+      </module>
+      <module
+            category="org.mwc.debrief.scripting.Debrief_Category"
+            class="org.mwc.debrief.scripting.wrappers.Spatial"
+            icon="icons/debrief_icon.png"
+            id="org.mwc.debrief.scripting.module"
+            name="Spatial"
+            visible="true">
+      </module>
+      <category
+            id="org.mwc.debrief.scripting.Debrief_Category"
+            name="Debrief">
+      </category>
    </extension>
 
 </plugin>
