+++ org.mwc.debrief.scripting/src/org/mwc/debrief/scripting/wrappers/DebriefObjects.java
@@ -47,51 +47,6 @@ public class DebriefObjects
     listenToMyParts();
   }
 
-  @Override
-  protected void finalize() throws Throwable
-  {
-    super.finalize();
-    System.out.println("DISPOSE");
-  }
-
-  public static WorldLocation createLocation(double dLat, double dLong,
-      double depth)
-  {
-    return new WorldLocation(dLat, dLong, depth);
-  }
-  
-  public WorldDistance createDistance(double km)
-  {
-    return new WorldDistance(km, WorldDistance.KM);
-  }
-  
-
-  public static ShapeWrapper createLine(WorldLocation startPt,
-      WorldLocation endPt, String name, Color color)
-  {
-    return new ShapeWrapper(name, new LineShape(startPt, endPt, name),
-        color, null);
-  }
-
-  public static ShapeWrapper createCircle(WorldLocation centre,
-      WorldDistance radius, String name, Color color)
-  {
-    return new ShapeWrapper(name, new CircleShape(centre, radius, name),
-        color, null);
-  }
-
-  public static WorldVector createVector(double distM, double bearingDegs)
-  {
-    return new WorldVector(MWC.Algorithms.Conversions.Degs2Rads(bearingDegs),
-        new WorldDistance(distM, WorldDistance.METRES), null);
-  }
-
-  public static WorldVector createVectorKm(double distKm, double bearingDegs)
-  {
-    return new WorldVector(MWC.Algorithms.Conversions.Degs2Rads(bearingDegs),
-        new WorldDistance(distKm, WorldDistance.KM), null);
-  }
-
   public static FixWrapper createFix(HiResDate time, WorldLocation location,
       double courseRads, double speedYps)
   {
@@ -184,7 +139,7 @@ public class DebriefObjects
       return null;
     }
 
-    public HiResDate getCurrentTime()
+    public HiResDate getTime()
     {
       TimeProvider time = (TimeProvider) _editor.getAdapter(TimeProvider.class);
       return time.getTime();
